nl80211: implement hapd_set_ssid callback
When operating in AP mode, this callback passes the SSID to the kernel
driver.

wl12xx cards handle probe requests in FW. When the SSID is not
broadcast in beacon data (i.e. the SSID is hidden), the AP must know the
real SSID in order to respond to directed probe requests.

Change-Id: I1b0ada51a8e9176957728229b6423349aec934b1
Signed-off-by: Arik Nemtsov <arik@wizery.com>
Signed-off-by: Vishal Mahaveer <a0271468@ti.com>